requirements.txt
# KisanAI - Stable Requirements
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# WebSocket and HTTP
websockets==12.0
python-multipart==0.0.6

# Machine Learning - Stable versions
torch
torchaudio
transformers==4.30.2
accelerate==0.20.3

# LangChain Ecosystem - Compatible versions
langchain==0.2.17
langchain-core>=0.2.27,<0.3.0
langchain-community>=0.2.19,<0.3.0
langgraph==0.2.16

# Vector Search and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Scientific Computing - Stable versions
numpy==1.24.3
scipy==1.10.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
ffmpeg-python==0.2.0
pyttsx3==2.90

# Language Detection
langid==1.1.6
langdetect==1.0.9

# Environment and Utilities
python-dotenv==1.0.0

# HuggingFace Hub - Compatible with sentence-transformers
huggingface-hub>=0.16.0,<0.20.0

# Additional dependencies
requests>=2.25.0
tqdm>=4.60.0
